FinePrint/requirements.txt
flask
google-genai
beautifulsoup4
requests
python-dotenv
openai
anthropic
